apps/ynab-transaction-reviewer/requirements.txt
# YNAB Transaction Reviewer Requirements

# Core dependencies
requests>=2.31.0
PyYAML>=6.0
python-dateutil>=2.8.2

# Google APIs for Gmail
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# AWS SDK (for Lambda deployment)
boto3>=1.28.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0